moving diffAnalyzer to a new file
